Added solutions

+++ LeetcodeGrind/Solutions/P0007.cs
@@ -5,7 +5,7 @@ public class P0007
 {
     public int Reverse(int x)
     {
-        double val = x % 10;
+        long val = x % 10;
 
         while (Math.Abs(x / 10) > 0)
         {
@@ -13,8 +13,7 @@ public class P0007
             val = (val * 10) + (x % 10);
         }
 
-        if (val < int.MinValue
-            || val > int.MaxValue)
+        if (val < int.MinValue || val > int.MaxValue)
         {
             return 0;
         }
